AWS Outpost Network CFN Template
This is the deliverable from the AWS Outpost lab week.  This template uses a Lambda Layer to create subnets with an exclusively set Availability Zone and Outpost ARN provided in the Template.  This allows for the pas-on-cloud CFN templates be installed on the AWS Outpost.  AWS requires that this deliverable be listed under the CyberArk GitHub account in order for our application to be accepted